feat: Add Derive function
To create "derived" sequences, such as the results of Map, add a helper
analogous to Generate that copies over the properties of the input stream
to the destination stream.

Include changes to Process, which backs all the map an filter operations
to use Derived. Also fix small bug in Process to use Each method instead
of top-level Each, so that early exit isn't masked there.