fix(async): add note to comment of Sync
Calling Sync on a asynchronous sequence may let you iterate over it
safely without needing to synchronize yourself, but it doesn't change
the fact that the values are coming from non-deterministic input, and
thus are still volatile.

Also, if the downstream code is already thread-safe, the coarse locking
in Sync may hurt performance.