refactor: make Repeat not use Limit
Repeat was previously a Limit() on Infinite(), which was fine, but had
two issues:
 - Slower, as we're adding a level of compute/indirection
 - Requires ErrStopIteration, which is a little buggy at the moment

Using Repeat(t,1) to implement Single is fine, as we're not adding
a level, and inlining should apply to limit overhead, but the buggy
end-condition will cause problems.

The bug with ErrStopIteration is that there is no distinction between
the case where a sequence is intentionally cut short by a wrapper that
is creating a new sequence, and the "Stop, I'm done" calls made inside
Each's callback to break out early. The former should not propogate to
upper levels while the latter should.