feat: update Simulate + tests
Update Simulate to take a sequence to add to. This allows 2 things:
 - Simulate can become a method
 - We can simulate multi-step algorithms with much less friction

The previous Simulate is now Simulation, and tests have been added
to ensure that Single(initial).Simulate(step) is logically equivalent
to Simulation(initial,step) and it simplifies the case where there
is only a single simulation step.