feat: add Result type
A Result is a (value,error) tuple that is meant to be returned when
functions would normally return (T,error), such that it can provide
utilities to handle errors, or outright ignore them should the situation
merit.

This is useful because sequence processing functions hide errors until
the final result is computed anyway. Also this tuple might get some use
in places where value/error is the norm.