refactor: introduce NextResult with no error
Make NextResult accept a callback that doesn't produce an error, such
that the only way an error makes it into the result is if the input had
an error. This should make it easier to use with conversion functions
that don't produce an error.

The old version is still available via NextResultErr.